backend/fastapi-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery==5.3.6
redis==5.0.1
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3